requirements.txt
numpy
opencv-python
matplotlib
tqdm
scipy
torchnet
